LayoutTests/imported/w3c/web-platform-tests/css/css-typed-om/the-stylepropertymap/properties/line-height-expected.txt


PASS Can set 'line-height' to CSS-wide keywords
FAIL Can set 'line-height' to var() references assert_equals: expected 2 but got 1
PASS Can set 'line-height' to the 'normal' keyword
FAIL Can set 'line-height' to a length assert_equals: expected "px" but got "em"
FAIL Can set 'line-height' to a number assert_equals: expected 2 but got 1
FAIL Can set 'line-height' to a percent assert_equals: expected 2 but got 1
PASS Setting 'line-height' to a time throws TypeError
PASS Setting 'line-height' to an angle throws TypeError
PASS Setting 'line-height' to a flexible length throws TypeError
PASS Setting 'line-height' to a URL throws TypeError
PASS Setting 'line-height' to a transform throws TypeError
